+++ nncf/common/factory.py
@@ -33,7 +33,7 @@ class NNCFGraphFactory:
         """
         model_backend = get_backend(model)
         if model_backend == BackendType.ONNX:
-            from nncf.experimental.onnx.graph.nncf_graph_builder import GraphConverter
+            from nncf.onnx.graph.nncf_graph_builder import GraphConverter
 
             return GraphConverter.create_nncf_graph(model)
         raise RuntimeError('Cannot create backend-specific graph'
@@ -50,7 +50,7 @@ class ModelTransformerFactory:
         """
         model_backend = get_backend(model)
         if model_backend == BackendType.ONNX:
-            from nncf.experimental.onnx.graph.model_transformer import ONNXModelTransformer
+            from nncf.onnx.graph.model_transformer import ONNXModelTransformer
             return ONNXModelTransformer(model)
         raise RuntimeError('Cannot create backend-specific model transformer'
                            'because {} is not supported!'.format(model_backend))
@@ -66,7 +66,7 @@ class EngineFactory:
         """
         model_backend = get_backend(model)
         if model_backend == BackendType.ONNX:
-            from nncf.experimental.onnx.engine import ONNXEngine
+            from nncf.onnx.engine import ONNXEngine
             return ONNXEngine(model)
         raise RuntimeError('Cannot create backend-specific engine'
                            'because {} is not supported!'.format(model_backend))
