+++ nncf/onnx/hardware/pattern_operations.py
@@ -13,24 +13,24 @@
 
 from nncf.common.graph.patterns import merge_two_types_of_operations
 from nncf.common.graph.graph_matching import GraphPattern
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXConvolutionMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXDepthwiseConvolutionMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXConvolutionTransposeMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXLinearMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXMatMulMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXBatchNormMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXReluMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXLeakyReluMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXThresholdedReluMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXEluMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXPReluMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXSigmoidMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXHardSigmoidMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXHardSwishMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXAddLayerMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXMulLayerMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXDivLayerMetatype
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import ONNXSubMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXConvolutionMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXDepthwiseConvolutionMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXConvolutionTransposeMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXLinearMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXMatMulMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXBatchNormMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXReluMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXLeakyReluMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXThresholdedReluMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXEluMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXPReluMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXSigmoidMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXHardSigmoidMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXHardSwishMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXAddLayerMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXMulLayerMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXDivLayerMetatype
+from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXSubMetatype
 
 LINEAR_OPERATIONS = {GraphPattern.METATYPE_ATTR: [ONNXConvolutionMetatype,
                                                   ONNXDepthwiseConvolutionMetatype,
