+++ tests/onnx/weightless_model.py
@@ -16,7 +16,7 @@ from copy import deepcopy
 import onnx
 from onnx import TensorProto
 from onnx.external_data_helper import uses_external_data
-from nncf.experimental.onnx.graph.onnx_graph import ONNXGraph
+from nncf.onnx.graph.onnx_graph import ONNXGraph
 import numpy as np
 import tempfile
